Removing SplashScreen and ConnectDialog adjustments
