+++ raiden-dapp/src/views/Home.vue
@@ -97,6 +97,7 @@ export default class Home extends Vue {
     await this.$raiden.connect(subkey);
     this.connectingSubkey = false;
     this.connecting = false;
+    this.connectDialog = false;
   }
 }
 </script>
